PropertyTreeのデフォルト値を設定

+++ tobas_hal/tobas_navio/tobas_navio_ros/include/tobas_navio_ros/rcin_handler.hpp
@@ -26,11 +26,11 @@ private:
   navio::RCInput rcin_;
 
   // Config
-  tobas_std::Range<int> roll_range_;
-  tobas_std::Range<int> pitch_range_;
-  tobas_std::Range<int> yaw_range_;
-  tobas_std::Range<int> thrust_range_;
-  std::array<int, tobas::kNumFlightModes> modes_;
+  tobas_std::Range<double> roll_range_;
+  tobas_std::Range<double> pitch_range_;
+  tobas_std::Range<double> yaw_range_;
+  tobas_std::Range<double> thrust_range_;
+  std::array<double, tobas::kNumFlightModes> modes_;
   double mode_auto_, mode_position_, mode__;
   double estop_on_, estop_off_;
   double gpsw_on_, gpsw_off_;
